Max depth is a dumb property now
